requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
langchain>=0.1.0
groq>=0.2.1     
python-dotenv>=1.0.0
